+++ Cargo.toml
@@ -35,7 +35,7 @@ clap = { version = "4.1.9", features = ["derive", "deprecated"] }
 clap_complete = "4.1.5"
 clap_mangen = "0.2.10"
 config = { version = "0.13.3", default-features = false, features = ["toml"] }
-criterion = "0.4.0"
+criterion = {version = "0.4.0", optional = true }
 crossterm = { version = "0.26", default-features = false }
 dirs = "4.0.0"
 esl01-renderdag = "0.3.0"
@@ -74,4 +74,5 @@ testutils = { path = "lib/testutils" }
 
 [features]
 default = ["jujutsu-lib/legacy-thrift"]
+bench = ["criterion"]
 vendored-openssl = ["git2/vendored-openssl", "jujutsu-lib/vendored-openssl"]
